made edit mode
